pymodel/requirements.txt
# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.24.0
pillow>=9.5.0
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 进度条
tqdm>=4.65.0

# 推理服务
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# GPU加速（可选，根据系统安装对应版本）
# 如需CUDA支持，请访问 https://pytorch.org/ 安装对应版本

# 注意：
# - Windows系统会自动使用 SimHei 字体
# - Linux/Mac系统如需中文显示，请安装中文字体或修改配置

